Move discovery view related classes to fragments.discover
