+++ babble/src/main/java/io/mosaicnetworks/babble/configure/TabsFragment.java
@@ -30,15 +30,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.google.android.material.tabs.TabLayout;
+import com.google.android.material.tabs.TabLayoutMediator;
+
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.viewpager2.adapter.FragmentStateAdapter;
 import androidx.viewpager2.widget.ViewPager2;
-
-import com.google.android.material.tabs.TabLayout;
-import com.google.android.material.tabs.TabLayoutMediator;
-
 import io.mosaicnetworks.babble.R;
+import io.mosaicnetworks.babble.fragments.discover.DiscoverGroupsFragment;
 
 /**
  * This Fragment implements a multi-tabbed UX allowing the display of multiple Discovery Fragments
